quantum-viz/pyproject.toml
[tool.poetry]
name = "quantum-viz"
version = "0.0.0"
description = "quantum-viz.js Python tools"
authors = ["Microsoft Corporation <que-contacts@microsoft.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/microsoft/quantum-viz.js"
repository = "https://github.com/microsoft/quantum-viz.js"
documentation = "https://github.com/microsoft/quantum-viz.js"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.urls]
Changelog = "https://github.com/microsoft/quantum-viz.js/releases"

[tool.poetry.dependencies]
python = "^3.6.1"
click = "^8.0.1"
ipython = "<8.0.0"
varname = "^0.8.1"
qiskit-terra = {extras = ["qiskit"], version = "^0.18.3"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
coverage = {extras = ["toml"], version = "^5.4"}
safety = "^1.10.3"
mypy = "^0.910"
typeguard = "^2.12.1"
xdoctest = {extras = ["colors"], version = "^0.15.5"}
sphinx = "^4.1.1"
sphinx-autobuild = "^2021.3.14"
pre-commit = "^2.13.0"
flake8 = "^3.9.2"
black = "^20.8b1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.4.3"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"
pep8-naming = "^0.12.0"
darglint = "^1.8.0"
reorder-python-imports = "^2.5.0"
pre-commit-hooks = "^4.0.1"
sphinx-rtd-theme = "^0.5.2"
sphinx-click = "^3.0.1"
Pygments = "^2.9.0"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["quantum_viz"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.mypy]
strict = false
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
